requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision 
torchaudio 


tensorflow
tensorboard
numpy
scipy
sklearn
matplotlib
opencv-python
opencv-contrib-python
mediapipe
termcolor


bosdyn-api                 
bosdyn-choreography-client 
bosdyn-choreography-protos 
bosdyn-client              
bosdyn-core                
bosdyn-mission             